refactor: moved all newly private UDP headers into private include folder
